+++ src/raw/cl/mod.rs
@@ -1,4 +1,4 @@
 pub mod abs;
 // pub mod enum_orgy;
-pub mod image;
+pub mod image_st;
 // pub mod flag;